TASK: Apply code migrations
TASK: Apply migration TYPO3.FLOW3-201201261636

Adjust to the major changes in FLOW3 1.1 when coming from 1.0.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Fluid-201205031303

Rename form.textbox to form.textfield

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.FLOW3-201205292145

Replace FileTypes use with MediaTypes use.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.FLOW3-201206271128

Replace DataNotSerializeableException with DataNotSerializableException.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.FLOW3-201209201112

Move all code to PSR-0 compatible directory structure, remove
Package.xml,

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-201209251426

Rename FLOW3 to TYPO3 Flow

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-201211151101

Check for globally defined role identifiers in Policy.yaml files

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-201212051340

Change entity resource definitions from using _ to \

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-201310031523

Change Neos\Flow\Persistence\Doctrine\DatabaseConnectionException to

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-201405111147

This migration does not actually change any code. It just displays a
warning if a TypeConverters canConvertFrom() is

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Fluid-20141113120800

Rename the "resource" argument of the security.ifAccess ViewHelper to
"privilegeTarget"

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-20141113121400

Adjust "Policy.yaml" to new syntax

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Fluid-20141121091700

Warn about removed ReflectionService dependency from AbstractViewHelper

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Fluid-20150214130800

Warn about "escapeOutput" property for existing ViewHelpers to ensure
backwards-compatibility

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-20151113161300

Adjust "Settings.yaml" to new "requestPattern" and "firewall" syntax
(see FLOW-412)

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-20161115140400

Adjust to the renaming of the Resource namespace and class in Flow 4.0

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.Flow-20161115140430

Adjust to the renaming of the Object namespace in Flow 4.0

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration Neos.Flow-20161124204700

Adjusts code to package renaming from "TYPO3.Flow" to "Neos.Flow"

TASK: Apply migration Neos.Flow-20161124204701

Adjusts code to package renaming from "Neos.Flow.Utility.Files" to
"Neos.Utility.Files" and other extractions of the "Utility" packages.

TASK: Apply migration Neos.Flow-20161124224015

Adjusts code to cache extraction

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration Neos.Eel-20161124230101

Adjusts code to Eel Renaming

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration Neos.Flow-20161125124112

Adjusts code to Neos\Flow\Utility\Unicode adjustment

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration TYPO3.FluidAdaptor-20161130112935

Adjusts code to package renaming from "TYPO3.Fluid" to
"Neos.FluidAdaptor".

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.